requirements.txt
python-telegram-bot==20.6
yt-dlp
python-dotenv
telegram
flask